+++ contrib/pg_upgrade/gp/test/integration/.gitignore
