+++ contrib/pg_upgrade/gp/test/integration/check_covering_aoindex_test.c
@@ -9,7 +9,7 @@
 #include "cmockery.h"
 
 #include "test_utils.h"
-#include "../checks.h"
+#include "../../checks.h"
 
 
 static void
